Update README.md
Removed a comma that I think was not necessary. Also changed "capable of making your programs up to 5 times faster" to "capable of making your programs considerably faster", since only some methods can be up to 5x faster and not the entire program necessarily. I also mention that on my contribution so in case they accept both of our contributions it would'nt be redundant.